+++ changelog.md
@@ -1,8 +1,19 @@
+# v1.3.0 - 02/12/2022
+
+Woolies API replaced the REST API endpoint for fetching transactions with a GraphQL API endpoint which provides similar
+but different information. This release adds support for this new endpoint. Note, Woolies will store transactions for
+only 14 months.
+
+- Added better support of Woolies' partners' e-receipts; i.e. BigW, BWS
+- Changed woolies transaction interface models
+- Fixed support for Woolies ✨new✨ API
+- Depreciated woolies-auth code 
+
 ## v1.2.0 - 10/04/2022
 
 - Added support for 2-up & multiplayer 💕 (multiple transactional accounts)
 - Added request retries & exponential back-off
-- Fixed incorrect comparison value; amount-paid now preferred over cost-of-items, to address gift-cards & discounts 
+- Fixed incorrect comparison value; amount-paid now preferred over cost-of-items, to address gift-cards & discounts
 - Fixed receipt-parsing for a few unseen edge-cases; i.e. price-reduced items
 
 ## v1.1.0 - 20/02/2022